21.11: Follow the pathsInNixDB → additionalPaths rename
This rename is unfortunate.  We depend here upon this actually being *all
the paths in the Nix DB*.  It still *is* all the paths in the Nix DB —
nothing has changed about its implementation — but now it's named after
one use case (adding some additional paths) rather than what it *is*
(all the paths), which makes this usage here read as if it couldn't
possibly work.  :(